+++ libv4l2/exynos_v4l2.c
@@ -140,7 +140,7 @@ int exynos_v4l2_open_devname(const char *devname, int oflag, ...)
             }
 
             /* read sysfs entry for device name */
-            size = (int)fgets(name, sizeof(name), stream_fd);
+            size = (intptr_t)fgets(name, sizeof(name), stream_fd);
             fclose(stream_fd);
 
             /* check read size */
